Add homepage components to home-page sub directory
