+++ package.json
@@ -15,6 +15,7 @@
     "react-intersection-observer": "^9.4.3",
     "react-router-dom": "^6.11.2",
     "react-scripts": "5.0.1",
+    "rough-notation": "^0.5.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
