+++ src/components/ContactPage.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-// import Header from './Header';
 
 const ContactPage = () => (
   <section className="form-page">
