+++ src/components/DraggableHeadshot.jsx
@@ -9,8 +9,7 @@
 //           isDragging: monitor.isDragging(),
 //         }),
 //       });
-      
-  
+
 //     return (
 //       <div ref={drag}>
 //         <img className="headshot" src={headshot} alt="headshot" style={{ opacity: isDragging ? 0.5 : 1 }} />
