+++ src/components/Home-page/HonourableMentions.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
 const HonourableMentions = () => (
-  <div className='honourable'>
-    <div className='honourable-div'>
+  <div className="honourable">
+    <div className="honourable-div">
       <a href="">APIs</a>
       <a href="">Machine Learning</a>
       <a href="">Block Chain</a>
