+++ src/styles/about-me.css
@@ -4,6 +4,7 @@
 .about-me-section {
   position: relative;
   margin: auto;
+
   /* background-color: #f0f0f0; */
 }
