+++ src/styles/honourable-mentions.css
@@ -22,7 +22,7 @@
 }
 
 .honourable-div a {
-  color:black;
+  color: black;
   font-family: 'Montserrat', sans-serif;
   padding: 20px;
   background-color: rgb(255, 246, 246);
