+++ src/styles/socials.css
@@ -4,7 +4,7 @@
   margin-top: 35px;
   border-radius: 50%;
   padding: 5px;
-  background: linear-gradient(to right, #fff 50%, #000000 50%);
+  background: linear-gradient(to right, #fff 50%, #000 50%);
   background-size: 200% 100%;
   background-position: right bottom;
   background-color: black;
@@ -19,10 +19,12 @@
 
 .socials-div {
   display: flex;
+
   /* width: 20%; */
   position: absolute;
   left: 90vw;
   top: 50vh;
+
   /* justify-content: flex-end; */
 }
